backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
nltk
pdfminer.six
python-multipart
# --- opcionais, se quiser IA externa ---
# transformers
# torch
# openai
